requirements.txt
# Core dependencies
requests>=2.28.0
PyYAML>=6.0

# Optional: For S3 upload support
boto3>=1.26.0

# Optional: For scheduled tasks
schedule>=1.2.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
responses>=0.23.0  # For mocking HTTP requests
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
